+++ Package.swift
@@ -45,6 +45,10 @@ let package = Package(
                 .copy("Foundation/GeoJSON/Fixtures/multipolygon.geojson"),
                 .copy("Helpers/MapboxAccessToken"),
                 .copy("Resources/empty-style-chicago.json"),
+                .copy("Resources/testShowsLogo().png"),
+                .copy("Resources/testDoNotShowsLogo().png"),
+                .copy("Resources/testShowsAttribution().png"),
+                .copy("Resources/testDoNotShowsAttribution().png"),
                 .process("Snapshot/Snapshot.xcassets"),
                 .process("Resources/MapInitOptionsTests.xib"),
             ]